+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/Amebis/eduVPN/compare/2.0.1...HEAD)
 
 - Launching client on sign-on is an opt-in now
+- Skip client re-authorization if expired on sign-on
 - .NET Framework 4.8 setup issue fixed